requirements.txt
Django==5.1.6
pillow==11.2.1
python-slugify==8.0.4
sqlparse==0.5.3
mysqlclient==2.2.7
